Update nl.json
I did some translations for Dutch,
not all of them, 
Sometimes there are multiple options to translate and choosing the correct words and syntax is a matter of preference. So if one has different preferences. I'm not claiming mine are the best.
I'm native Dutch, however my HA installation is fully in English. So I will probably never see those translations myself.