Update README.md
Added an instruction about restarting an appliance that does not show up in the integration while showing up in the mobile app without wifi. I had this issue and was about to open an issue when I found other users that faced the same issue and fixed it by restarting the appliance, and that worked for me too.